+++ src/main/lua/lrpureraw.lrdevplugin/Info.lua
@@ -22,6 +22,6 @@ return {
         id = "noRejected", -- unique identifier for export filter
     },
 
-    VERSION = { major = 1, minor = 0, revision = 2, build = 3, },
+    VERSION = { major = 1, minor = 1, revision = 0, build = 0, },
 
 }