+++ CHANGELOG.md
@@ -45,3 +45,4 @@ Only the documentation has changed. No need to update.
 ### Fixed
 Fix installation chapter in README.md 
 
+